requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
bleach=5.0.0=pypi_0
blis=0.7.7=pypi_0
brotlipy=0.7.0=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.3.29=h06a4308_0
catalogue=2.0.7=pypi_0
certifi=2020.6.20=pyhd3eb1b0_3
cffi=1.15.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=pypi_0
cryptography=36.0.0=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cymem=2.0.6=pypi_0
debugpy=1.6.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
entrypoints=0.4=pypi_0
executing=0.8.3=pypi_0
fastjsonschema=2.15.3=pypi_0
ffmpeg=4.3=hf484d3e_0
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
idna=3.3=pyhd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.13.0=pypi_0
ipython=8.2.0=pypi_0
ipython-genutils=0.2.0=pypi_0
it-core-news-lg=3.2.0=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.1=pypi_0
jpeg=9d=h7f8727e_0
jsonschema=4.4.0=pypi_0
jupyter-client=7.2.2=pypi_0
jupyter-core=4.9.2=pypi_0
jupyterlab-pygments=0.2.1=pypi_0
lame=3.100=h7b6447c_0
langcodes=3.3.0=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.1.1=pypi_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-fft=1.3.1=pypi_0
mkl-random=1.2.2=pypi_0
mkl-service=2.4.0=pypi_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
murmurhash=1.0.6=pypi_0
nbclient=0.6.0=pypi_0
nbconvert=6.5.0=pypi_0
nbformat=5.3.0=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.5=pypi_0
nettle=3.7.3=hbbd107a_1
notebook=6.4.10=pypi_0
numpy=1.21.2=pypi_0
numpy-base=1.21.2=py310h79a1101_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1n=h7f8727e_0
packaging=21.3=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathy=0.6.1=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=pypi_0
pip=21.2.4=pypi_0
preshed=3.0.6=pypi_0
prometheus-client=0.14.1=pypi_0
prompt-toolkit=3.0.29=pypi_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.8.2=pypi_0
pygments=2.11.2=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.7=pypi_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=pypi_0
python=3.10.4=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytorch=1.11.0=py3.10_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pyyaml=6.0=pypi_0
pyzmq=22.3.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
send2trash=1.8.0=pypi_0
setuptools=58.0.4=pypi_0
six=1.16.0=pyhd3eb1b0_1
smart-open=5.2.1=pypi_0
soupsieve=2.3.2=pypi_0
spacy=3.2.4=pypi_0
spacy-legacy=3.0.9=pypi_0
spacy-loggers=1.0.2=pypi_0
sqlite=3.38.2=hc218d9a_0
srsly=2.4.2=pypi_0
stack-data=0.2.0=pypi_0
terminado=0.13.3=pypi_0
thinc=8.0.15=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.1.1=pypi_0
tk=8.6.11=h1ccaba5_0
torch=1.11.0=pypi_0
torchaudio=0.11.0=pypi_0
torchvision=0.12.0=pypi_0
tornado=6.1=pypi_0
tqdm=4.64.0=pypi_0
traitlets=5.1.1=pypi_0
typer=0.4.1=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.8=pyhd3eb1b0_0
wasabi=0.9.1=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
